gradio/requirements (1).txt
torch==2.0
gradio
midi2audio
pydub
matplotlib
torchsummary
scikit-learn
einops
pyfluidsynth
mido
Pillow